overview fixes
