+++ src/app/gallery/_components/FullImageView.tsx
