+++ src/app/gallery/_components/ImageComponent.tsx
@@ -1,4 +1,5 @@
 "use client";
+
 import Image from "next/image";
 import type { ImageType } from "~/types/ImageType";
